Fixed gm_sidereal() so it will not generate warning -Werror=long-long when compiled with -Wextra
git-svn-id: https://svn.savannah.nongnu.org/svn/avr-libc/trunk/avr-libc@2334 1084037a-409f-4a99-8160-1a4b3577d4c1

+++ libc/time/gm_sidereal.c
@@ -35,11 +35,12 @@
 */
 
 #include <time.h>
+#include <stdint.h>
 
 unsigned long
 gm_sidereal(const time_t * timer)
 {
-    unsigned long long tmp;
+    uint64_t tmp;
 
     tmp = *timer;
 
@@ -50,7 +51,7 @@ gm_sidereal(const time_t * timer)
     tmp >>= 31;
 
     /* add sidereal time at epoch */
-    tmp += 23991ULL;
+    tmp += (uint64_t)23991;
 
     tmp %= ONE_DAY;
     return tmp;
